third_party/WebKit/LayoutTests/fast/css/usecounter-pseudo-any-link.html

<!DOCTYPE html>
<html>
  <head>
    <title>Use counter test for :any-link</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <style id="style"></style>
    <a href="https://www.google.com" id="link">Link</a>
    <script>
      'use strict';
      test(() => {
        let CSSSelectorPseudoAnyLink = 2264;
        let isCounted = () => internals.isUseCounted(document, CSSSelectorPseudoAnyLink);
        var style = document.getElementById('style');
        var div = document.createElement('div');

        style.textContent = ':-webkit-any-link { color: green; }';
        getComputedStyle(link);
        assert_false(isCounted(), ':any-link should not be counted');

        style.textContent = ':any-link { color: blue; }';
        assert_equals(getComputedStyle(link).color, 'rgb(0, 0, 255)');
        assert_true(isCounted(), ':any-link should be counted');
      }, ':any-link counter is working');
    </script>
  </head>
<html>
